Add checking if QPixmap object is Null
There are images which filename extensions are not their actual formats
for some reason so the app cannot create a proper QPixmap object (it
becomes Null). In this case we try to create the object passing specific
image formats. If this does not work either, create an object with the
error image