Remove option 'Caching thumbnails'
There is no point of it. If we cache thumbnails of the duplicate images the majority of which
we are going to delete later anyway, there is no point. If we cache thumbnails (and create)
of the all images, again, most of them is going to be never used (usually, the rate of the duplicate
images is small, and people run these kind of programmes quite rarely).